rm OpenCV dependecy

+++ include/conversion.h
@@ -17,7 +17,7 @@ typedef struct Conversion* Conversion_t;
 
 Conversion_t Init_Conversion();
 
-int H264_to_BGR24(Conversion_t conv, uint8_t** dst, uint8_t* src, int size);
+int H264_to_BGR24(Conversion_t conv, uint8_t** dst, uint8_t* src, size_t size);
 
 void Free_Conversion(Conversion_t* conv);
